requirements.txt
git+https://github.com/Rapptz/discord.py
nekos.life-async
python-dotenv
waifuim.py
alexflipnote.py
animec
psutil
asyncdagpi
humanize
asyncpraw
googlesearch-python